requirements.txt
Flask==1.1.1
gunicorn==19.7.1
python-dotenv==0.13.0
pymongo==3.10.1